Fix part of #4044: Prepare for supporting math expressions (math utility refactor) (#4046)
## Explanation
Fix part of #4044
Originally copied from #2173 when it was in proof-of-concept form

This PR introduces a new math.proto and math utility package for upcoming protos & utilities coming in later PRs as part of the broader math expressions project. This includes moving some pure math protos & utilities to the new package, but not quite everything can be moved yet due to resource dependencies for RatioExtensions and the ratio/fraction parsers. These could be split up and partly moved to the new package in a future PR, but it's not a prerequisite for any of the math expressions work.

This PR also moves most of fraction parsing to utility and extracts just the UI-specific portion (i.e. accessing app strings). Consequently, there are now two test suites to verify the fraction parsing functionality. StringExtensions needed to be moved out of app due to it being needed in FractionParser. This move is needed for a downstream PR that will utilize the fraction parser in tests.

The new test file exemption changes are just updates since there are moved files.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If this PR fixes part of an issue, prefix the title with "Fix part of #bugnum: ...".)
- [x] Any changes to [scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets) files have their rationale included in the PR explanation.
- [x] The PR follows the [style guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android Studio ([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers ([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A refactor with no side effects to affect user flows.

Commit history:

* Copy proto-based changes from #2173.

* Introduce math.proto & refactor math extensions.

Much of this is copied from #2173.

* Migrate tests & remove unneeded prefix.

* Add needed newline.

* Some needed Fraction changes.

* Lint fix.

* Fix broken test post-refactor.

* Post-merge fix.

* Add regex check, docs, and resolve TODOs.

This also changes regex handling in the check to be more generic for
better flexibility when matching files.

* Lint fix.

* Fix failing static checks.

* Move StringExtensions & fraction parsing.

This splits fraction parsing between UI & utility components.

* Add missing KDocs.

* Fix broken build.

* Fix broken build post-merge.

* Post-merge fix.

* More post-merge fixes.

+++ domain/src/main/java/org/oppia/android/domain/classify/rules/fractioninput/FractionInputIsEquivalentToAndInSimplestFormRuleClassifierProvider.kt
@@ -6,9 +6,9 @@ import org.oppia.android.app.model.WrittenTranslationContext
 import org.oppia.android.domain.classify.RuleClassifier
 import org.oppia.android.domain.classify.rules.GenericRuleClassifier
 import org.oppia.android.domain.classify.rules.RuleClassifierProvider
-import org.oppia.android.domain.util.approximatelyEquals
-import org.oppia.android.domain.util.toFloat
-import org.oppia.android.domain.util.toSimplestForm
+import org.oppia.android.util.math.approximatelyEquals
+import org.oppia.android.util.math.toDouble
+import org.oppia.android.util.math.toSimplestForm
 import javax.inject.Inject
 
 /**
@@ -36,6 +36,7 @@ class FractionInputIsEquivalentToAndInSimplestFormRuleClassifierProvider
     input: Fraction,
     writtenTranslationContext: WrittenTranslationContext
   ): Boolean {
-    return answer.toFloat().approximatelyEquals(input.toFloat()) && answer == input.toSimplestForm()
+    return answer.toDouble().approximatelyEquals(input.toDouble()) &&
+      answer == input.toSimplestForm()
   }
 }
